Update docs for gui entry
